+++ tests/testthat/test-get_rd.R
@@ -7,14 +7,15 @@ test_that("While searching for relevant R documentation, Errors return silently"
 })
 
 test_that("Can find relevant R documentation files, given a .Rd filepath", {
-  expect_s3_class(get_rd(file = file.path(test_path(), "data" , "man", "autovalidate.Rd")), "Rd")
+  expect_s3_class(get_rd(file = rd_filepaths[[1L]]), "Rd")
 })
 
 test_that("get_rd find macros properly", {
-  expect_s3_class(get_rd(file = system.file("examples", "rd_file_sample_2.Rd", package = "rd2markdown"), macros = NA), "Rd")
-  rd_file <- get_rd(file = system.file("examples", "rd_file_sample_2.Rd", package = "rd2markdown"), macros = NA)
+  rd_filepath <- system.file("examples", "rd_file_sample_2.Rd", package = "rd2markdown")
+  expect_s3_class(get_rd(file = rd_filepath, macros = NA), "Rd")
+  expect_silent(rd <- get_rd(file = rd_filepath, macros = NA))
   expect_equivalent(
-    rd_file[[17]][[3]], 
+    rd[[17]][[3]],
     structure(c("\\code{\\link[=#1]{#1()}}", "rnorm"), Rd_tag = "USERMACRO", macro = "\\function")
   )
 })